+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-26 12:01:06
-// Commit: test(core): add unit tests for core business logic
+// Updated at: 2026-08-26 15:33:26
+// Commit: fix(core): prevent double tap duplicate action in core
 
 class coreComponentState {
-  final String commitRef = '3933';
-  final String lastMessage = 'test(core): add unit tests for core business logic';
-  final DateTime timestamp = DateTime.parse('2026-08-26T12:01:06+05:00');
+  final String commitRef = '3947';
+  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
+  final DateTime timestamp = DateTime.parse('2026-08-26T15:33:26+05:00');
 }